requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-secretmanager>=2.16.0
python-gitlab>=3.15.0
requests>=2.31.0
pyyaml>=6.0
gitpython>=3.1.0
jinja2>=3.1.0
black>=23.0.0
pytest>=7.4.0
flake8>=6.0.0 